Localization/Code/requirements.txt
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==2.1.1
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pycountry==22.3.5
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
smmap==5.0.1
urllib3==1.26.18
